% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{add_label}
\alias{add_label}
\title{Add a column of SNP labels to input data.frame}
\usage{
add_label(merged, snp)
}
\arguments{
\item{merged}{(data.frame) Input data.frame, which is a result by merging two
association studies. See the example under `get_lead_snp()` for generation of
such data.frame.}

\item{snp}{(character vector) A vector of SNP rsIDs. If only labeling one SNP,
this can also be a single string.}
}
\description{
Add a column of SNP labels to input data.frame
}
\examples{
# Select the lead SNP
in_fn_1 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn_1, marker_col = 'rsid', pval_col = 'pval')
in_fn_2 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d2 = read_metal(in_fn_2, marker_col = 'rsid', pval_col = 'pval')
merged = merge(d1, d2, by = "rsid", suffixes = c("1", "2"), all = FALSE)
merged = add_label(merged, 'rs9349379')
}
