% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{assign_color}
\alias{assign_color}
\title{Assign color to each SNP according to LD.}
\usage{
assign_color(rsid, snp, ld)
}
\arguments{
\item{rsid}{(character vector) A vector of rsIDs on which to assign color.}

\item{snp}{(string) rsID for lead SNP. This SNP will be colored purple.
Other SNPs will be assigned color based on their LD with the lead SNP.}

\item{ld}{(data.frame) The output from `retrieve_LD()`.}
}
\description{
Assign color to each SNP according to LD.
}
\examples{
# Select the lead SNP
in_fn_1 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn_1, marker_col = 'rsid', pval_col = 'pval')
in_fn_2 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d2 = read_metal(in_fn_2, marker_col = 'rsid', pval_col = 'pval')
merged = merge(d1, d2, by = "rsid", suffixes = c("1", "2"), all = FALSE)
ld <- retrieve_LD('6', 'rs9349379', 'AFR')
color = assign_color(rsid = merged$rsid, snp = 'rs9349379', ld)
}
