% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config}
\alias{config}
\title{Example dataset}
\format{
## `who`
A data frame with 1 row and 3 columns:
\describe{
  \item{a}{"URL"}
  \item{b}{"locuscomparer"}
  \item{c}{"12345678"}
}
}
\usage{
config
}
\description{
Example dataset
}
\keyword{datasets}
