% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{locuscompare3}
\alias{locuscompare3}
\title{Make a three way locuscompare plot.}
\usage{
locuscompare3(
  in_fn1,
  in_fn2,
  in_fn3,
  marker_col1 = "rsid",
  pval_col1 = "pval",
  title1 = "GWAS",
  marker_col2 = "rsid",
  pval_col2 = "pval",
  title2 = "caQTL",
  marker_col3 = "rsid",
  pval_col3 = "pval",
  title3 = "eQTL",
  snp = NULL,
  population = "EUR",
  combine = TRUE,
  legend = TRUE,
  legend_position = c("bottomright", "topright", "topleft"),
  lz_ylab_linebreak = FALSE,
  genome = c("hg19", "hg38")
)
}
\arguments{
\item{in_fn1}{(string) Path to the input file for study 1.}

\item{in_fn2}{(string) Path to the input file for study 3.}

\item{marker_col1}{(string, optional) Name of the marker column. Default: 'rsid'.}

\item{pval_col1}{(string, optional) Name of the p-value column. Default: 'pval'.}

\item{title1}{(string) The title for the x-axis.}

\item{marker_col2}{(string, optional) Name of the marker column. Default: 'rsid'.}

\item{pval_col2}{(string, optional) Name of the p-value column. Default: 'pval'.}

\item{title2}{(string) The title for the y-axis.}

\item{marker_col3}{(string, optional) Name of the marker column. Default: 'rsid'.}

\item{pval_col3}{(string, optional) Name of the p-value column. Default: 'pval'.}

\item{title3}{(string) The title for the y-axis.}

\item{snp}{(string, optional) SNP rsID. If NULL, the function will select the lead SNP. Default: NULL.}

\item{population}{(string, optional) One of the 5 popuations from 1000 Genomes: 'AFR', 'AMR', 'EAS', 'EUR', and 'SAS'. Default: 'EUR'.}

\item{combine}{(boolean, optional) Should the three plots be combined into one plot? If FALSE, a list of
three plots will be returned. Default: TRUE.}

\item{legend}{(boolean, optional) Should the legend be shown? Default: TRUE.}

\item{legend_position}{(string, optional) Either 'bottomright','topright', or 'topleft'. Default: 'bottomright'.}

\item{lz_ylab_linebreak}{(boolean, optional) Whether to break the line of y-axis of the locuszoom plot.}

\item{genome}{(string, optional) Genome assembly, either 'hg19' or 'hg38'. Default: 'hg19'.}
}
\description{
Make a three way locuscompare plot.
}
\examples{
in_fn1 = system.file('extdata','gwas.tsv', package = 'locuscomparer')
in_fn2 = system.file('extdata','eqtl.tsv', package = 'locuscomparer')
locuscompare(in_fn1 = in_fn1, in_fn2 = in_fn2, snp="rs12032")
}
