% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Make a scatter plot (called the LocusCompare plot).
Each axis of the LocusCompare plot represent the -log10(p-value) from
an association study. Each point thus represent a SNP. By default, the lead SNP
is a purple diamond, whereas the other SNPs are colored according to
their LD with the lead SNP.}
\usage{
make_scatterplot(
  merged,
  title1,
  title2,
  color,
  shape,
  size,
  legend = TRUE,
  legend_position = c("bottomright", "topright", "topleft")
)
}
\arguments{
\item{merged}{(data.frame) An input data.frame which has the following
columns: rsid, pval1 (p-value for study 1), logp1 (p-value for study 2),
logp1 (log p-value for study 1), logp2 (log p-value for study 2), chr, pos.
See the example for `get_lead_snp()` on how to generate this data.frame.}

\item{title1}{(string) The title for the x-axis.}

\item{title2}{(string) The title for the y-axis.}

\item{color}{(data.frame) The output from `assign_color()`.}

\item{shape}{(data.frame) Specification of the shape of each SNP. See example blow on how to generate this data.frame.}

\item{size}{(data.frame) Specification of the size of each SNP. See example below on how to generate this data.frame.}

\item{legend}{(boolean) Whether to include the legend.}

\item{legend_position}{(string, optional) Either 'bottomright','topright', or 'topleft'. Default: 'bottomright'.}
}
\description{
Make a scatter plot (called the LocusCompare plot).
Each axis of the LocusCompare plot represent the -log10(p-value) from
an association study. Each point thus represent a SNP. By default, the lead SNP
is a purple diamond, whereas the other SNPs are colored according to
their LD with the lead SNP.
}
\examples{
# Select the lead SNP
in_fn_1 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn_1, marker_col = 'rsid', pval_col = 'pval')
in_fn_2 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d2 = read_metal(in_fn_2, marker_col = 'rsid', pval_col = 'pval')
merged = merge(d1, d2, by = "rsid", suffixes = c("1", "2"), all = FALSE)
merged = add_label(merged, 'rs9349379')
ld <- retrieve_LD('6', 'rs9349379', 'AFR')
color = assign_color(rsid = merged$rsid, snp = 'rs9349379', ld)
snp = 'rs9349379'
shape = ifelse(merged$rsid == snp, 23, 21)
names(shape) = merged$rsid
size = ifelse(merged$rsid == snp, 3, 2)
names(size) = merged$rsid
make_scatterplot(merged, title1 = 'GWAS', title2 = 'eQTL', color, shape, size)
}
