% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{read_metal}
\alias{read_metal}
\title{Read association summary statistics and append normalized columns.}
\usage{
read_metal(in_fn, chromosome_col = "chromosome", position_col = "position",
  logp_col = "-log10 p-value")
}
\arguments{
\item{in_fn}{(string) Path to the input file.}

\item{chromosome_col}{(string, optional) Name of the chromosome column. Default: 'chromosome'.}

\item{position_col}{(string, optional) Name of the position column. Default: 'position'.}

\item{logp_col}{(string, optional) Name of the -log10(p-value) column. Default: '-log10 p-value'.}
}
\description{
Read association summary statistics and append normalized columns.
}
\examples{
in_fn = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn, chromosome_col = 'chromosome', position_col = 'position', logp_col = '-log10 p-value')
}
