% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{add_label}
\alias{add_label}
\title{Add a column of SNP labels to input data.frame}
\usage{
add_label(merged, snp)
}
\arguments{
\item{merged}{(data.frame) Input data.frame, which is a result by merging two
association studies. See the example under `get_lead_snp()` for generation of
such data.frame.}

\item{snp}{(character vector) A vector of SNP rsIDs. If only labeling one SNP,
this can also be a single string.}
}
\description{
Add a column of SNP labels to input data.frame
}
\examples{
# The data.frame merged comes from the example for `get_lead_snp()`.
merged = add_label(merged, 'rs9349379')
}
