% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config}
\alias{config}
\title{information on AWS server for locuscompare}
\format{
simple data.frame
}
\usage{
config
}
\description{
information on AWS server for locuscompare
}
\keyword{datasets}
