% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{get_position}
\alias{get_position}
\title{Append two columns, chromosome (chr) and position (pos), to the input data.frame.}
\usage{
get_position(x, genome = c("hg19", "hg38"))
}
\arguments{
\item{x}{(data.frame) Input data.frame.}

\item{genome}{(string, optional) Genome assembly, either 'hg19' or 'hg38'. Default: 'hg19'.}
}
\description{
Append two columns, chromosome (chr) and position (pos), to the input data.frame.
}
\examples{
in_fn = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn, marker_col = 'rsid', pval_col = 'pval')
genome = "hg19"
get_position(d1, genome)
}
