% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{make_locuszoom}
\alias{make_locuszoom}
\title{Make a locuszoom plot.
Details see http://locuszoom.org/.}
\usage{
make_locuszoom(metal, title, chr, color, shape, size, ylab_linebreak = FALSE)
}
\arguments{
\item{metal}{(data.frame) input file with two column, rsID and p-value.
See `read_metal()` for more details.}

\item{title}{(string) y-axis title.}

\item{chr}{(string) chromosome.}

\item{color}{(data.frame) The output from `assign_color()`.}

\item{shape}{(data.frame) Specification of the shape of each SNP. See example blow on how to generate this data.frame.}

\item{size}{(data.frame) Specification of the size of each SNP. See example below on how to generate this data.frame.}

\item{ylab_linebreak}{(boolean, optional) Whether to break the line of y-axis. If FALSE, the y-axis title and '-log10(p-value)'
will be on the same line. Default: FALSE.}
}
\description{
Make a locuszoom plot.
Details see http://locuszoom.org/.
}
\examples{
in_fn_1 = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn_1, marker_col = 'rsid', pval_col = 'pval')
in_fn_2 = system.file('extdata', 'eqtl.tsv', package = 'locuscomparer')
d2 = read_metal(in_fn_2, marker_col = 'rsid', pval_col = 'pval')
merged = merge(d1, d2, by = "rsid", suffixes = c("1", "2"), all = FALSE)
snp = 'rs9349379'
merged = add_label(merged, snp)
d1 = add_label(d1, snp)
shape = ifelse(merged$rsid == snp, 23, 21)
names(shape) = merged$rsid
size = ifelse(merged$rsid == snp, 3, 2)
names(size) = merged$rsid
chr = '6'
d1 = get_position(d1)
ld = retrieve_LD(chr, snp, "AFR")
color = assign_color(rsid = merged$rsid, snp = 'rs9349379', ld)
make_locuszoom(d1, title = 'GWAS', chr, color, shape, size)
}
