% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuscompare.R
\name{read_metal}
\alias{read_metal}
\title{Read association summary statistics from file and append column.
The file must contain 2 columns: markers, i.e SNPs, and p-value.
The marker column should contain SNP rsIDs.}
\usage{
read_metal(in_fn, marker_col = "rsid", pval_col = "pval")
}
\arguments{
\item{in_fn}{(string) Path to the input file.}

\item{marker_col}{(string, optional) Name of the marker column. Default: 'rsid'.}

\item{pval_col}{(string, optional) Name of the p-value column. Default: 'pval'.}
}
\description{
Read association summary statistics from file and append column.
The file must contain 2 columns: markers, i.e SNPs, and p-value.
The marker column should contain SNP rsIDs.
}
\examples{
in_fn = system.file('extdata', 'gwas.tsv', package = 'locuscomparer')
d1 = read_metal(in_fn, marker_col = 'rsid', pval_col = 'pval')
}
